Move query human output helpers into partial
